+++ tvm_assembler/src/simple.rs
@@ -839,6 +839,7 @@ impl Engine {
         DIFF_PATCH_BINARY_ZIPQ               => 0xC7, 0x25
         MINTECC                              => 0xC7, 0x26
         CNVRTSHELLQ                          => 0xC7, 0x27
+        MINTSHELL                            => 0xC7, 0x28
         CALCBKREWARD                         => 0xC7, 0x29
         CALCMINSTAKE                         => 0xC7, 0x30
     }