+++ tvm_block/src/types.rs
@@ -112,7 +112,7 @@ macro_rules! define_VarIntegerN {
             // Interface to write value with type rule
             fn write_to_cell(value: &BigInt) -> Result<BuilderData> {
                 let len = Self::get_len(value);
-                if len >= $N {
+                if len > $N {
                     fail!("serialization of {} error {} >= {}", stringify!($varname), len, $N)
                 }
 
@@ -847,6 +847,11 @@ impl CurrencyCollection {
         Self::from_grams(Grams::from(grams))
     }
 
+    pub fn set_grams(&mut self, grams: u64) -> Result<()> {
+        self.grams = Grams::from(grams);
+        Ok(())
+    }
+
     pub const fn from_grams(grams: Grams) -> Self {
         CurrencyCollection { grams, other: ExtraCurrencyCollection::default() }
     }
@@ -997,6 +1002,13 @@ impl Serializable for u32 {
 impl Serializable for u128 {
     fn write_to(&self, cell: &mut BuilderData) -> Result<()> {
         cell.append_u128(*self)?;
+        Ok(())
+    }
+}
+
+impl Serializable for i128 {
+    fn write_to(&self, cell: &mut BuilderData) -> Result<()> {
+        cell.append_i128(*self)?;
         Ok(())
     }
 }