+++ tvm_executor/src/ordinary_transaction.rs
@@ -82,6 +82,7 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
         in_msg: Option<&Message>,
         account: &mut Account,
         params: ExecuteParams,
+        minted_shell: &mut u128,
     ) -> Result<Transaction> {
         #[cfg(feature = "timings")]
         let mut now = Instant::now();
@@ -125,7 +126,7 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
         let mut acc_balance = account.balance().cloned().unwrap_or_default();
         let mut msg_balance = in_msg.get_value().cloned().unwrap_or_default();
         let gas_config = self.config().get_gas_config(false);
-        log::debug!(target: "executor", "src_dapp_id = {:?}, address = {:?}", params.src_dapp_id, in_msg.int_header());
+        log::debug!(target: "executor", "src_dapp_id = {:?}, address = {:?}, available_credit {:?}", params.src_dapp_id, in_msg.int_header(), params.available_credit);
         if let Some(h) = in_msg.int_header() {
             if params.src_dapp_id != account.get_dapp_id().cloned()
                 && !(in_msg.have_state_init()
@@ -343,6 +344,8 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
                         new_data,
                         account_address,
                         is_special,
+                        params.available_credit,
+                        minted_shell,
                     ) {
                         Ok(ActionPhaseResult { phase, messages, copyleft_reward }) => {
                             out_msgs = messages;