+++ tvm_vm/src/executor/engine/handlers.rs
@@ -387,6 +387,7 @@ impl Handlers {
                 .set(0x23, execute_diff_patch_binary_zip_not_quiet)
                 .set(0x24, execute_diff_patch_binary_quiet)
                 .set(0x25, execute_diff_patch_binary_zip_quiet)
+                .set(0x28, execute_mint_shell)
                 .set(0x26, execute_ecc_mint)
                 .set(0x27, execute_exchange_shell)
                 .set(0x29, execute_calculate_validator_reward)